+++ mayan/apps/document_states/models/workflow_instance_models.py
@@ -31,22 +31,26 @@ class WorkflowInstance(
     _ordering_fields = ('datetime',)
 
     workflow = models.ForeignKey(
-        on_delete=models.CASCADE, related_name='instances', to=Workflow,
+        help_text=_('The workflow template.'), on_delete=models.CASCADE,
+        related_name='instances', to=Workflow,
         verbose_name=_(message='Workflow')
     )
     datetime = models.DateTimeField(
-        auto_now_add=True, db_index=True, help_text=_(
-            'Workflow instance creation date time.'
-        ), verbose_name=_(message='Datetime')
+        auto_now_add=True, db_index=True,
+        help_text=_('Workflow instance creation date time.'),
+        verbose_name=_(message='Datetime')
     )
     document = models.ForeignKey(
-        on_delete=models.CASCADE, related_name='workflows', to=Document,
+        help_text=_(
+            'The document to which the workflow instance is attached.'
+        ), on_delete=models.CASCADE, related_name='workflows', to=Document,
         verbose_name=_(message='Document')
     )
     context = models.TextField(
         blank=True, verbose_name=_(message='Context')
     )
     state_active = models.ForeignKey(
+        help_text=_('The currently active state of the workflow instance.'),
         on_delete=models.CASCADE, related_name='workflow_instances',
         to=WorkflowState, verbose_name=_(message='Active state')
     )