+++ HISTORY.rst
@@ -177,6 +177,8 @@
   - Use `select` class for the inclusion and operator fields.
 
 - Make source metadata available for linking.
+- Add help texts to the workflow instance fields: workflow, document,
+  state active.
 
 4.8.3 (2024-10-04)
 ==================